+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 5.1.1 - Unreleased
 
-- Confirmed support for CockroachDB 24.2.x (no code changes required).
+- Confirmed support for CockroachDB 24.2.x and 24.3.x (no code changes
+  required).
 
 ## 5.1 - 2024-08-13