+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 5.2 - Unreleased
+## 5.2 - 2025-04-07
 
 Initial release for Django 5.2.x and CockroachDB 23.2.x, 24.1.x, 24.3.x, and
 25.1.x.